He terminado los viajes con su logica de negocio.pero falta pdf y pruebas god, VIP pruebas

+++ src/main/kotlin/com/es/jwtSecurityKotlin/service/DestinoService.kt
@@ -1,13 +1,12 @@
 ﻿package com.es.jwtSecurityKotlin.service
 
+import com.es.jwtSecurityKotlin.exception.ConflictoBD
+import com.es.jwtSecurityKotlin.exception.NotFoundException
 import com.es.jwtSecurityKotlin.model.Destino
 import com.es.jwtSecurityKotlin.repository.DestinoRepository
-import com.example.unsecuredseguros.exception.ConflictoBD
-import com.example.unsecuredseguros.exception.NotFoundException
+import com.es.jwtSecurityKotlin.repository.ViajeRepository
 import org.springframework.security.core.Authentication
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.data.jpa.domain.AbstractPersistable_.id
-import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
 
 @Service
@@ -15,6 +14,9 @@ class DestinoService {
 
     @Autowired
     private lateinit var destinoRepository: DestinoRepository
+
+    @Autowired
+    private lateinit var viajeRepository: ViajeRepository
 
     //TODO Canbiar null por errores personalizados
 
@@ -26,8 +28,8 @@ class DestinoService {
 
         val destinos = destinoRepository.findAll()
 
-        if (destinos.any { it.origen.equals(destino.origen) }) {
-            throw ConflictoBD("El destino ${destino.origen} existente")
+        if (destinos.any { it.nombre.uppercase().equals(destino.nombre.uppercase()) }) {
+            throw ConflictoBD("El destino ${destino.nombre} existente")
         }
 
         destinoRepository.save(destino)
@@ -36,7 +38,13 @@ class DestinoService {
 
     fun putDestino(destino: Destino , idDestino:Long): Destino {
         val destinobd = destinoRepository.findById(idDestino)
-            .orElseThrow { NotFoundException("No se ha encontrado el destino a actualizar.")}
+            .orElseThrow { NotFoundException("No se ha encontrado el destino a actualizar.") }
+        //No puedo poner 2 destinos iguales
+        val destinos = destinoRepository.findAll()
+
+        if (destinos.any { it.nombre.uppercase().equals(destino.nombre.uppercase()) }) {
+            throw ConflictoBD("El destino ${destino.nombre} existente")
+        }
 
         destinobd.origen = destino.origen
         destinobd.pais = destino.pais
@@ -51,6 +59,12 @@ class DestinoService {
     fun deleteDestino(idDestino: Long): String{
 
         val destino = destinoRepository.findById(idDestino).orElseThrow { NotFoundException("No se ha encontrado el destino a eliminar.") }
+
+        //Hay que mirar si existe en la tabla de viajes, sino existe se puede eliminar con tranquilidad.
+        if ( viajeRepository.findAll().any { it.destino.idDestino?.equals(destino.idDestino) == true }) {
+            throw ConflictoBD("El destino no puede ser eliminado ya que está asociado a 1 o más viajes.")
+        }
+
         destinoRepository.delete(destino)
         return "Se ha eliminado correctamente el destino."
     }
